# backend/Swagger Documentation.yaml
openapi: 3.0.3
info:
  title: Swagger SmartHoover - OpenAPI 3.0
  description: |-
    Documentation for the SmartHoover integrated application was developed by UkrainGroup

  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: http://localhost:8080/api
...
paths:
  /salesman/:
    get:
      tags:
        - salesman
      summary: Get all salesmen
      description: Get all salesmen
      operationId: getAllSalesmen
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Salesman'
        '404':
          description: Salesman not found
        '422':
          description: Validation exception
        '500':
          description: Server error
          #security:
          # - petstore_auth:
          #    - write:pets
          #   - read:pets

    post:
      tags:
        - salesman
      summary: Add a new pet to the store
      description: Add a new pet to the store
      operationId: addPet
      requestBody:
        description: Create a new pet in the store
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Pet'
          application/xml:
            schema:
              $ref: '#/components/schemas/Pet'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Pet'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Pet'
            application/xml:
              schema:
                $ref: '#/components/schemas/Pet'
        '400':
          description: Invalid input
        '422':
          description: Validation exception
      security:
        - petstore_auth:
            - write:pets
            - read:pets
...
components:
  schemas:
    Salesman:
      required:
        - firstName
        - lastName
        - employeeId
        - code
      type: object
      properties:
        firstName:
          type: string
          example: John
        middleName:
          type: string
          example: Ilya
        lastName:
          type: string
          example: Smith
        fullName:
          type: string
          example: John Ilya Smith
        employeeId:
          type: integer
          format: int64
          example: 12421
        code:
          type: integer
          format: int64
          example: 12421
      xml:
        name: salesman
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    api_key:
      type: apiKey
      name: api_key
      in: header